// src/Moryx.FactoryMonitor.Web/app/src/app/api/models/factory-state-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActivityChangedModel } from '../models/activity-changed-model';
import { CellLocationModel } from '../models/cell-location-model';
import { CellStateChangedModel } from '../models/cell-state-changed-model';
import { OrderModel } from '../models/order-model';
import { ResourceChangedModel } from '../models/resource-changed-model';
export interface FactoryStateModel {
  activityChangedModels?: Array<ActivityChangedModel> | null;
  backgroundURL?: string | null;
  cellStateChangedModels?: Array<CellStateChangedModel> | null;
  hasManufacturingFactory?: boolean;
  iconName?: string | null;
  id?: number;
  isACell?: boolean;
  location?: CellLocationModel;
  orderModels?: Array<OrderModel> | null;
  resourceChangedModels?: Array<ResourceChangedModel> | null;
}