// src/Moryx.FactoryMonitor.Web/app/src/app/api/models/resource-changed-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CellLocationModel } from '../models/cell-location-model';
import { CellPropertySettings } from '../models/cell-property-settings';
export interface ResourceChangedModel {
  cellIconName?: string | null;
  cellImageURL?: string | null;
  cellLocation?: CellLocationModel;
  cellName?: string | null;
  cellPropertySettings?: ({
[key: string]: CellPropertySettings;
}) | null;
  factoryId?: number;
  iconName?: string | null;
  id?: number;
  isACell?: boolean;
  location?: CellLocationModel;
}