// src/Moryx.FactoryMonitor.Web/app/src/app/api/models/activity-changed-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActivityClassification } from '../models/activity-classification';
import { OrderReferenceModel } from '../models/order-reference-model';
export interface ActivityChangedModel {
  classification?: ActivityClassification;
  id?: number;
  orderReferenceModel?: OrderReferenceModel;
  resourceId?: number;
}